changed: removed unused addon types, added ADDON_UNKNOWN & ADDON_SCRAPER_LIBRARY types fixed: compiler warnings
removed: more unused ADDON_TYPE s
fixed: no need for ADDON:: namespace qualifier in AddonManager.cpp

git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/addons-fwork@27223 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/GUIDialogAddonBrowser.cpp
@@ -133,7 +133,7 @@ bool CGUIDialogAddonBrowser::OnMessage(CGUIMessage& message)
 
   case GUI_MSG_SETFOCUS:
     {
-      if (m_viewControl.HasControl(message.GetControlId()) && (DWORD) m_viewControl.GetCurrentControl() != message.GetControlId())
+      if (m_viewControl.HasControl(message.GetControlId()) && m_viewControl.GetCurrentControl() != message.GetControlId())
       {
         m_viewControl.SetFocused();
         return true;
